REFACTOR: Moved messages and tests of bank module to own file.
